Project refactoring

+++ src/main/java/com/camperfire/marketflow/repository/InventoryRepository.java
@@ -3,5 +3,5 @@ package com.camperfire.marketflow.repository;
 import com.camperfire.marketflow.model.Inventory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InventoryRepository extends JpaRepository<Inventory, Integer> {
+public interface InventoryRepository extends JpaRepository<Inventory, Long> {
 }
